INC-68 fix: fix bags, update styles, refactor
