+++ features/back/Back.tsx
@@ -1,25 +1,34 @@
 import { ArrowLeft } from "@/public";
 import { Button, Typography } from "@/shared/components";
-import Link from "next/link";
+import { clsx } from "clsx";
 import { useRouter } from "next/router";
 
 import s from "./back.module.scss";
 
 type Props = {
+  className?: string;
   text: string;
 };
-export const Back = ({ text }: Props) => {
+export const Back = ({ className, text }: Props) => {
   const router = useRouter();
   const handleGoBack = () => {
     router.back();
   };
 
+  const linkCN = clsx(s.link, className);
+
   return (
-    <div className={s.root}>
-      <Button className={s.link} onClick={handleGoBack} variant={"icon"}>
-        <ArrowLeft />
-        <Typography variant={"regular_text-14"}>{text}</Typography>
-      </Button>
-    </div>
+    <Button
+      aria-label={text}
+      className={linkCN}
+      onClick={handleGoBack}
+      variant={"icon"}
+    >
+      <ArrowLeft />
+
+      <Typography className={s.text} variant={"regular_text-14"}>
+        {text}
+      </Typography>
+    </Button>
   );
 };
