+++ features/back/back.module.scss
@@ -1,20 +1,16 @@
-@import "../../styles/mixins";
-
-.root {
-  width: 100%;
-  @include flex("", center, center, 12px);
-}
-
 .link {
-  padding: 0 30px;
-  max-width: 1100px;
-  width: 100%;
-  margin: 0 auto;
-  @include flex("", flex-start, center, 12px);
   &:hover {
     color: var(--color-accent-100);
   }
+
   &:active {
     color: var(--color-accent-500) !important;
+  }
+}
+
+.text {
+  padding-left: 6px;
+  @media screen and (max-width: 576px) {
+    display: none;
   }
 }
