+++ pages/auth/policy/index.tsx
@@ -1,7 +1,6 @@
 import { Back } from "@/features";
-import { useTranslation } from "@/shared/assets/hooks/useTranslation";
-import { PageWrapper, Typography } from "@/shared/components";
-import { HeadMeta } from "@/shared/components/headMeta/HeadMeta";
+import { useTranslation } from "@/shared/assets/hooks";
+import { HeadMeta, Typography } from "@/shared/components";
 import { getLayout } from "@/shared/components/layout/baseLayout/BaseLayout";
 
 import s from "./policy.module.scss";
@@ -10,13 +9,13 @@ const PrivacyPolicy = () => {
   const { t } = useTranslation();
 
   return (
-    <PageWrapper>
+    <>
       <HeadMeta title={"Privacy Policy"} />
-      <Back text={"Back to Sign Up"} />
+      <Back className={s.back} text={"Back to Sign Up"} />
       <Typography className={s.title} variant={"h1"}>
         {t.policy.title}
       </Typography>
-      <Typography className={s.text} variant={"regular_text-14"}>
+      <Typography as={"p"} className={s.text} variant={"regular_text-14"}>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
         tempor incididunt ut labore et dolore magna aliqua. Fames ac turpis
         egestas integer eget aliquet nibh. Amet consectetur adipiscing elit ut
@@ -93,7 +92,7 @@ const PrivacyPolicy = () => {
         volutpat est velit egestas dui id ornare. Feugiat nibh sed pulvinar
         proin.
       </Typography>
-    </PageWrapper>
+    </>
   );
 };
