+++ pages/auth/policy/policy.module.scss
@@ -1,10 +1,21 @@
+.back {
+  @media screen and (max-width: 576px) {
+    position: fixed;
+  }
+}
+
 .title {
-  margin-top: 20px;
+  margin: 20px auto;
+  text-align: center;
+  @media screen and (max-width: 576px) {
+    max-width: 80%;
+    margin: 4px auto 10px;
+  }
 }
 
 .text {
   max-width: 950px;
   width: 100%;
   text-align: center;
-  margin-top: 20px;
+  margin: 0 auto;
 }
