+++ shared/components/card/card.module.scss
@@ -15,9 +15,3 @@
     1px 1px 2px rgba(0 0 0 / 10%),
     -1px -1px 2px rgba(0 0 0 / 10%);
 }
-
-@media screen and (width <= 640px) {
-  .card {
-    width: 320px;
-  }
-}
