+++ shared/components/layout/layout.module.scss
@@ -1,8 +1,9 @@
 .root {
   width: 100%;
   height:100vh;
+  padding: 0 40px;
   @media screen and (max-width: 576px) {
-    padding: 15px;
+    padding: 15px 12px;
     justify-content: normal;
   }
 }
@@ -15,6 +16,9 @@
   width: 100%;
   max-width: 1920px;
   margin: 60px auto 0;
+  @media screen and (max-width: 576px) {
+    padding: 0;
+  }
 
   &::-webkit-scrollbar {
     //background-color: transparent;
