+++ shared/components/pageWrapper/pageWrapper.module.scss
@@ -1,6 +1,9 @@
 .root {
   position: relative;
   margin-left: 220px;
+  border: 2px solid red;
+  height: 100%;
+  min-height: calc(100vh - 100px);
   @media screen and (max-width: 768px) {
     margin-left: 0;
     margin-bottom: 150px;
