+++ shared/components/typography/typography.module.scss
@@ -4,7 +4,6 @@
   font-size: $fontSize;
   font-weight: $fontWeight;
   line-height: $lineHeight;
-  color: var(--color-text-primary);
 }
 
 .Large {
