+++ widgets/header/header.module.scss
@@ -9,10 +9,13 @@
 
   width: 100%;
   height: 60px;
-  padding: 12px;
+  padding: 12px 60px;
 
   background-color: var(--color-dark-700);
   border-bottom: 1px solid var(--color-dark-300);
+  @media screen and (max-width: 576px) {
+    padding: 12px 14px;
+  }
 }
 
 .header {
@@ -21,12 +24,9 @@
   width: 100%;
   max-width: 1920px;
   margin: 0 auto;
-  padding: 0 35px;
 
   color: var(--color-light-100);
-  @media screen and (max-width: 576px) {
-    padding: 0;
-  }
+
 }
 
 .logo {
