+++ widgets/profile-photo/edit-profilePhoto.module.scss
@@ -14,7 +14,7 @@
 
 .wrapperButton {
   display: flex;
-  justify-content: end;
+  justify-content: flex-end;
   margin-top: 26px;
 }
